requirements.txt
streamlit
numpy
pickle-mixin
tensorflow
keras
requests
pillow
opencv-python-headless
opencv-python
matplotlib